refactor chevron link styles into own file
