+++ src/components/molecules/chevronLink/chevronLink.styles.ts
@@ -1,0 +1,10 @@
+import { CSSProperties } from 'react'
+
+export const chevronLinkStyles: { [key: string]: CSSProperties } = {
+  wrapper: { display: 'flex', alignItems: 'center' },
+  chevron: {
+    marginLeft: 7,
+    width: 11,
+    height: 11,
+  },
+}
